Add diffusion
It's cool to compare AA and non-AA because the reason non-AA doesn't
work and our diffusion does, is because AA is already sampling 100
times.

So for each point on the sphere, we're actually getting 100 samples from
the environment in a random direction from that point, and that's what
averaged together gives us our diffusion effect.